requirements.txt
discord.py
toml
jishaku